project/model/requirements.txt
torch
flask==2.0.1
werkzeug==2.0.1
six
google-auth
google-cloud-aiplatform